requirements.txt
paho-mqtt==1.6.1
ollama==0.6.0